Force big-endian byte order
This tweak is necessary to read back Yaffs2 file systems dumped from
big-endian devices on little-endian systems.  Another requirement is
that Yaffs2 source code must be patched to handle tags ECC properly.
Unfortunately, Yaffs1 code is affected by the same bug in a different
function, where it is not as easy to fix as for Yaffs2, nor is it likely
that it ever will be as Yaffs1 code is currently pretty much set in
stone.  Given that I also do not have access to a real-world device
using Yaffs1, a patch for this deficiency was never prepared and Yafut
is currently only able to read back Yaffs2 file systems in an
endian-portable manner; therefore, '.stored_endian = 2' should not be
used when testing Yaffs1 compatibility with nandsim.